requirements.txt
Django==4.2.7
psycopg2-binary>=2.9.9
python-decouple==3.8
requests==2.31.0
google-generativeai==0.3.1
Pillow>=10.0.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7

